+++ packages/typir/src/kinds/class/class-initializer.ts
@@ -4,13 +4,13 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import { Type, TypeStateListener } from '../../graph/type-node.js';
+import { isType, Type, TypeStateListener } from '../../graph/type-node.js';
 import { TypeInitializer } from '../../initialization/type-initializer.js';
 import { InferenceProblem, InferenceRuleNotApplicable, TypeInferenceRule } from '../../services/inference.js';
 import { TypirServices } from '../../typir.js';
-import { InferenceRuleWithOptions, optionsBoundToType, bindInferCurrentTypeRule, ValidationRuleWithOptions, bindValidateCurrentTypeRule } from '../../utils/utils-definitions.js';
-import { MapListConverter, checkNameTypesMap, createTypeCheckStrategy } from '../../utils/utils-type-comparison.js';
-import { assertType } from '../../utils/utils.js';
+import { bindInferCurrentTypeRule, bindValidateCurrentTypeRule, InferenceRuleWithOptions, optionsBoundToType, ValidationRuleWithOptions } from '../../utils/utils-definitions.js';
+import { checkNameTypesMap, createTypeCheckStrategy, MapListConverter } from '../../utils/utils-type-comparison.js';
+import { assertType, toArray } from '../../utils/utils.js';
 import { ClassKind, CreateClassTypeDetails, InferClassLiteral } from './class-kind.js';
 import { ClassType, isClassType } from './class-type.js';
 
@@ -34,8 +34,9 @@ export class ClassTypeInitializer<LanguageType = unknown> extends TypeInitialize
         }
 
         this.createInferenceAndValidationRules(this.typeDetails, this.initialClassType);
-        // register all the inference rules already now to enable early type inference for this Class type
-        this.inferenceRules.forEach(rule => services.Inference.addInferenceRule(rule.rule, optionsBoundToType(rule.options, undefined))); // 'undefined', since the Identifier is still missing
+        // register all the inference rules already now to enable early type inference for this Class type ('undefined', since the Identifier is still missing)
+        this.inferenceRules.forEach(rule => services.Inference.addInferenceRule(rule.rule, optionsBoundToType(rule.options, undefined)));
+        this.validationRules.forEach(rule => services.validation.Collector.addValidationRule(rule.rule, optionsBoundToType(rule.options, undefined)));
 
         this.initialClassType.addListener(this, true); // trigger directly, if some initialization states are already reached!
     }
@@ -63,16 +64,22 @@ export class ClassTypeInitializer<LanguageType = unknown> extends TypeInitialize
 
             // remove the inference rules for the invalid type
             this.inferenceRules.forEach(rule => this.services.Inference.removeInferenceRule(rule.rule, optionsBoundToType(rule.options, undefined)));
+            this.validationRules.forEach(rule => this.services.validation.Collector.removeValidationRule(rule.rule, optionsBoundToType(rule.options, undefined)));
             // but re-create the inference rules for the new type!!
             // This is required, since inference rules for different declarations in the AST might be different, but should infer the same Typir type!
             this.createInferenceAndValidationRules(this.typeDetails, readyClassType);
+            // add the new rules
             this.inferenceRules.forEach(rule => this.services.Inference.addInferenceRule(rule.rule, optionsBoundToType(rule.options, readyClassType)));
+            this.validationRules.forEach(rule => this.services.validation.Collector.addValidationRule(rule.rule, optionsBoundToType(rule.options, readyClassType)));
         } else {
             // the class type is unchanged (this is the usual case)
 
             // keep the existing inference rules, but register it for the unchanged class type
             this.inferenceRules.forEach(rule => this.services.Inference.removeInferenceRule(rule.rule, optionsBoundToType(rule.options, undefined)));
+            this.validationRules.forEach(rule => this.services.validation.Collector.removeValidationRule(rule.rule, optionsBoundToType(rule.options, undefined)));
+
             this.inferenceRules.forEach(rule => this.services.Inference.addInferenceRule(rule.rule, optionsBoundToType(rule.options, readyClassType)));
+            this.validationRules.forEach(rule => this.services.validation.Collector.addValidationRule(rule.rule, optionsBoundToType(rule.options, readyClassType)));
         }
     }
 
@@ -114,7 +121,10 @@ export class ClassTypeInitializer<LanguageType = unknown> extends TypeInitialize
         }
         for (const inferenceRulesForClassLiterals of typeDetails.inferenceRulesForClassLiterals) {
             this.inferenceRules.push(this.createInferenceRuleForLiteral(inferenceRulesForClassLiterals, classType));
-            // TODO validation
+            const validationRule = this.createValidationRuleForLiteral(inferenceRulesForClassLiterals, classType);
+            if (validationRule) {
+                this.validationRules.push(validationRule);
+            }
         }
         for (const inferenceRulesForFieldAccess of typeDetails.inferenceRulesForFieldAccess) {
             this.inferenceRules.push({
@@ -151,7 +161,35 @@ export class ClassTypeInitializer<LanguageType = unknown> extends TypeInitialize
                     // boundToType: ... this property will be specified outside of this method
                 },
             });
-            // TODO validation
+            const validationRules = toArray(inferenceRulesForFieldAccess.validation);
+            if (validationRules.length >= 1) {
+                this.validationRules.push({
+                    rule: (languageNode, accept, typir) => {
+                        if (inferenceRulesForFieldAccess.filter !== undefined && inferenceRulesForFieldAccess.filter(languageNode) === false) {
+                            return;
+                        }
+                        if (inferenceRulesForFieldAccess.matching !== undefined && inferenceRulesForFieldAccess.matching(languageNode) === false) {
+                            return;
+                        }
+                        const field = inferenceRulesForFieldAccess.field(languageNode);
+                        if (field === InferenceRuleNotApplicable) {
+                            return;
+                        }
+                        const fieldType = typeof field === 'string'
+                            ? classType.getFields(true).get(field)
+                            : typir.Inference.inferType(field);
+                        if (isType(fieldType) === false) {
+                            return;
+                        }
+                        // TODO review: insert 'fieldType' as additional parameter?
+                        validationRules.forEach(rule => rule(languageNode, classType, accept, typir));
+                    },
+                    options: {
+                        languageKey: inferenceRulesForFieldAccess.languageKey,
+                        // boundToType: ... this property will be specified outside of this method
+                    },
+                });
+            }
         }
     }
 
@@ -169,7 +207,8 @@ export class ClassTypeInitializer<LanguageType = unknown> extends TypeInitialize
                             if (inputArguments.size >= 1) {
                                 return mapListConverter.toList(inputArguments);
                             } else {
-                                // there are no operands to check
+                                // skip this step for nominally typed classes
+                                // TODO this needs to be reworked for structural classes!
                                 return classType; // this case occurs only, if the current class has no fields (including fields of super types) or is nominally typed
                             }
                         } else {
@@ -200,10 +239,57 @@ export class ClassTypeInitializer<LanguageType = unknown> extends TypeInitialize
                             subProblems: checkedFieldsProblems,
                         };
                     } else {
-                        // the current function is not overloaded, therefore, the types of their parameters are not required => save time, ignore inference errors
                         return classType;
                     }
                 },
+            },
+            options: {
+                languageKey: rule.languageKey,
+                // boundToType: ... this property will be specified outside of this method
+            },
+        };
+    }
+
+    protected createValidationRuleForLiteral<T extends LanguageType>(rule: InferClassLiteral<LanguageType, T>, classType: ClassType): ValidationRuleWithOptions<LanguageType, T> | undefined {
+        const validationRules = toArray(rule.validation);
+        if (validationRules.length <= 0) {
+            return undefined;
+        }
+        return {
+            rule: (languageNode, accept, typir) => {
+                if (rule.filter !== undefined && rule.filter(languageNode) === false) {
+                    return;
+                }
+                if (rule.matching !== undefined && rule.matching(languageNode) === false) {
+                    return;
+                }
+                const inputArguments = rule.inputValuesForFields(languageNode);
+                if (inputArguments.size >= 1) {
+                    // check the given arguments
+                    const allExpectedFields = classType.getFields(true);
+                    if (allExpectedFields.size !== inputArguments.size) {
+                        return;
+                    }
+                    const compareFieldTypes = createTypeCheckStrategy(this.kind.options.subtypeFieldChecking, typir);
+                    for (const [fieldName, argumentValue] of inputArguments.entries()) {
+                        const actualFieldType = typir.Inference.inferType(argumentValue);
+                        if (isType(actualFieldType) === false) {
+                            return; // inference problem => skip validations
+                        }
+                        const expectedFieldType = allExpectedFields.get(fieldName);
+                        if (expectedFieldType === undefined) {
+                            return; // argument for a non-existing parameter
+                        }
+                        if (compareFieldTypes(actualFieldType, expectedFieldType) !== undefined) {
+                            return; // types are different
+                        }
+                        // everything is fine with this argument
+                    }
+                } else {
+                    // skip this step for nominally typed classes
+                    // TODO this needs to be reworked for structural classes!
+                }
+                validationRules.forEach(rule => rule(languageNode, classType, accept, typir));
             },
             options: {
                 languageKey: rule.languageKey,